changed folders structure
