+++ docs/src/components/layouts/Main.tsx
@@ -1,0 +1,38 @@
+import {
+	AdaptivityProvider,
+	AppRoot,
+	ConfigProvider,
+	Panel,
+	SplitCol,
+	SplitLayout,
+	View,
+	WebviewType,
+} from '@vkontakte/vkui';
+import React, {FC} from 'react';
+
+import Header from './shared/Header';
+
+type Props = {
+	children: React.ReactNode;
+};
+
+const Main: FC<Props> = ({children}) => (
+	<ConfigProvider webviewType={WebviewType.INTERNAL}>
+		<AdaptivityProvider>
+			<AppRoot>
+				<SplitLayout>
+					<SplitCol>
+						<View activePanel="main">
+							<Panel id="main">
+								<Header />
+								<div className="container">{children}</div>
+							</Panel>
+						</View>
+					</SplitCol>
+				</SplitLayout>
+			</AppRoot>
+		</AdaptivityProvider>
+	</ConfigProvider>
+);
+
+export default Main;
