+++ docs/src/components/layouts/shared/Header.content.ts
@@ -1,0 +1,26 @@
+type MenuItem = {
+	text: string;
+	href?: string;
+};
+
+export const menuItems: Array<MenuItem> = [
+	{
+		text: 'Документация',
+		href: 'https://vkcom.github.io/VKUI/',
+	},
+	{
+		text: 'Компоненты',
+		href: 'https://vkcom.github.io/VKUI/#/SplitLayout',
+	},
+	{
+		text: 'Токены',
+	},
+	{
+		text: 'Иконки',
+		href: 'https://vkcom.github.io/icons/',
+	},
+	{
+		text: 'Github',
+		href: 'https://github.com/VKCOM/VKUI',
+	},
+];
