+++ docs/src/components/layouts/shared/Header.tsx
@@ -1,34 +1,8 @@
 import {PanelHeader, Tappable, Text} from '@vkontakte/vkui';
 import React, {FC} from 'react';
 
-import {LogoIcon} from '../shared/content/icons';
-
-type MenuItem = {
-	text: string;
-	href?: string;
-};
-
-const menuItems: Array<MenuItem> = [
-	{
-		text: 'Документация',
-		href: 'https://vkcom.github.io/VKUI/',
-	},
-	{
-		text: 'Компоненты',
-		href: 'https://vkcom.github.io/VKUI/#/SplitLayout',
-	},
-	{
-		text: 'Токены',
-	},
-	{
-		text: 'Иконки',
-		href: 'https://vkcom.github.io/icons/',
-	},
-	{
-		text: 'Github',
-		href: 'https://github.com/VKCOM/VKUI',
-	},
-];
+import {LogoIcon} from '../../../shared/content/icons';
+import {menuItems} from './Header.content';
 
 const styles = {
 	tappable: {
