+++ docs/src/components/pages/Tokens/index.ts
@@ -1,0 +1,2 @@
+export {default as TokensActions} from './TokensActions/TokensActions';
+export {default as TokensHeader} from './TokensHeader';
