+++ docs/src/pages/Tokens.tsx
@@ -1,7 +1,6 @@
 import React, {FC} from 'react';
 
-import TokensActions from './TokensActions/TokensActions';
-import TokensHeader from './TokensHeader';
+import {TokensActions, TokensHeader} from '../components/pages/Tokens';
 
 const Tokens: FC = () => (
 	<div className="space-y-24px">
