+++ docs/src/styles/flex.css
@@ -1,0 +1,11 @@
+.flex {
+	display: flex;
+}
+
+.items-center {
+	align-items: center;
+}
+
+.justify-between {
+	justify-content: space-between;
+}
