+++ docs/src/styles/space.css
@@ -1,0 +1,11 @@
+.space-x-1px > *:not(:first-child) {
+	margin-left: 1px;
+}
+
+.space-x-20px > *:not(:first-child) {
+	margin-left: 20px;
+}
+
+.space-y-24px > *:not(:first-child) {
+	margin-top: 24px;
+}
